Convert MFSL/MFSL_ASYNC pool_dirs and pool_files to new management macros
This conversion attempts to preserve the intent of the old code. I'm not
sure the code was correct in the first place. Prealloced entries are
never freed, further, the code does some initialization on the entries
after the pool is created, but this initialization won't happen if the
pool is expanded, further, with _NO_BLOCK_PREALLOC, there won't
be any entries to initialize.

Signed-off-by: Frank Filz <ffilzlnx@us.ibm.com>