Remove many conditional compiles based on _USE_PROXY
The places _USE_PROXY is tested to enforce version 4 only have been removed in favor of using the macro to define the default NFS version to NFS v4 only, and to allow the NFS_Protocols option in the core params to be v4 only.

Also, force the protocol versions allowed in Exports to be a subset of the versions in the core params.

Also, disable NFS v2 by default.

Signed-off-by: Frank Filz <ffilzlnx@us.ibm.com>