XFS fsal refactor
Refactor out WRAP_* functions and cast typedefs internal to the fsal.
As with the VFS refactor, there are now common functions now that the
fsal specific typedefs have been pushed lower.

NOTE: This work has uncovered a problem in MainNFSD/nfs_file_content_flush_thread.c
which breaks dynamic multiple fsal support work if XFS is one of the
included filesystems.  It is patched and noted for now but export contexts
need rework.

Signed-off-by: Jim Lieb <jlieb@panasas.com>