replace creden_t with global user_credentials struct in ZFS FSAL
Pushing the creden_t structure down into the FSAL at the point
of the zfs library call and using user_credentials everywhere
above that makes a whole lot of fsal code common.  Now that it
is gone, ZFS looks a lot more like the other fsals.  This is the
first step in refactoring the fsal_op_context structure into a
common part and a fsal private part.

Signed-off-by: Jim Lieb <jlieb@panasas.com>