Do not instantiate Data Server handles in Cache_inode
Data server handles do not support metadata operations and a PUTFH on
one should not require contacting the metadata server to populate the
cache.  Additionally, in a mixed metadata/data server, a data server
handle may compare as equal to a metadata handle for the same file,
resulting in routing or other information not being passed to the FSAL
DS functions.  Extracting the FSAL handle specifically on DS calls
avoids this.