Race condition between worker & fridge threads causes the TIRPC XPRT object to return error & close the connection in addition, RPC handler thread isnt serialized.
the issue seems to be a race condition between the fridge thread & the worker thread.
the fridge is reading the RPC off the socket, selects a worker thread & then call DispatchWorkNFS() which:
    (*) pushes the RPC into that workers pending_request queue
    (*) signals a condition variable "workers_data[worker_index].wcb.tcb_condvar" - this will triger the worker to start working.
once this function returns (back to process_rpc_request()), it will immediately measure the time it took the RPC to execute WITHOUT waiting for any signal from the worker that it had actually completed (or even started) the execution.
The fridge is then back on the XPRT object, attempting to read the next request off the socket.
the worker is then executing the RPC, writing the reply.
at that point, the fridge will wake up with an error from the XPRT & indicate that the connection DIED.

The above scenario also doesnt make sure to serialize RPC's of a single stream, thereby it might return incorrect replies in case the client would post multiple RPC's without waiting for replies (as they might be dependent on one another)
i've fixed this with an additional condvar in the request_data_t object which is signaled by the worker once the RPC processing was completed & waited by the fridge to prevent it from trying to read the next RPC off the socket.