Refactor cache_entry_t
Consolidate fsal_handle and attributes to outside the object union.
This simplifies a lot of code, both no longer required switch statements
but long dereference chains in expressions.

This simplifies changes to this structure that will be part of the
new api changes.

Fix missing lock in cache_inode_setattr

Signed-off-by: Jim Lieb <jlieb@panasas.com>