Update handle transformation and validation functions
* More rigorous protocol handle validation.
* Update transforms between internal and protocol handles both for
  "real" handles and pseudo handles.
* DUILD_BUG_ON is used to compile time validate handle sizes.
* Handles are now passed around with fsal_handle_desc.

Other than now having a version number which makes the protocol more
robust, it also uses the descriptor so sizes can be validated all the
way down to the FSAL layer.  This is also the base for managing the
handle hash keys because we *know* exactly how big a handle really is
at all levels.

Signed-off-by: Jim Lieb <jlieb@panasas.com>

Update nfs_file_handle_mgmt.c xform

squash

Signed-off-by: Jim Lieb <jlieb@panasas.com>