Restructure hash table and cache inode
This is a major change to how the inode cache works.  Prior to this
change, it worked as follows:

Get/lookup:
* allocate space for a key (the handle bits)
* copy the key opaque bits (the whole opaque) into that space
* call a hash table function with that key
* free the key and continue on your way.

New entry:
* do a get as above.
* If missed, allocate space for a key and fill it as above
* allocate a cache entry and init it, including copying the handle
* call the FSAL with that handle to get info for the cache entry
* call the hash table function to insert it into the table
* free the space used for the "get"

On errors the code has to free the cache entry, the hash key for it, and
the key used for the get/lookup.

A free of a cache entry also must free the key from the hash table.

In the new way we have a pointer and length to the handle either in
the protocol payload or the cache entry :

Get/Lookup:
* Create a key by pointing directly to the handle opaque contents in the
  the protocol buffer.
* call the hash table function with that key.

There is nothing to free on either hit or miss.  The descriptor used to
compose the key is on the stack.

New entry:
* do a get as above
* If missed, allocate a cache entry, call the FSAL, and fill in the struct.
* copy the opaque handle part from the protocol buffer into the handle
  in the cache entry.
* fill in the handle descriptor in the cache entry to point to the handle in
  the cache entry.
* construct a key from the descriptor
* call the hash table function to insert the entry.

A free of  the cache entry (once removed from the hash) requires no free
of a key because the key is the handle element.

This change eliminates all memory allocations other than the cache entry
itself.

Signed-off-by: Jim Lieb <jlieb@panasas.com>