XFS FSAL - only use bytes which have been filled in by libhandle
We save the length of the handle provided by libhandle when
we copy it into the xfsfsal_handle_t and we can use it to
make sure we only ship useful bits around.

Also expand inode to 64 bits to cater for larger XFS filesystems.

NB! For non-shared FSAL builts one cannot use zero length array
for handle_val - xfsfsal_handle_t is placed on the stack in
various places.

Signed-off-by: Max Matveev<makc@gmx.co.uk>
Signed-off-by: Jim Lieb <jlieb@panasas.com>