XFS FSAL - deal with non-regular files
Symlinks are not the only filesystem objects which cannot be
opened using libhandle's open_by_handle interface: sockets
fifos, block and character device are in the same boat. In
order to deal with them switch from "open first, stat second"
to "stat first and only open what can be opened" mode.

And use real types when constructing file handles.
Regular files and directories are left with DT_UNKNOWN
to preserve on-the-wire format of the file handles.

Signed-off-by: Max Matveev <makc@gmx.co.uk>