Renew and Initialize NFSv4.1 clientids
Recent work to implement the NFSv4 lease timer modified the
SETCLIENTID and RENEW operations to update the time of the most recent
use of the clientid.  This updates the clientid time in EXCHANGE_ID
and SEQUENCE which was previously 0, cause all NFSv4.1 clientids to
report as expired.  This partially resolves the bug reported by Marc
Eshel on non-pNFS reads from NFSv4.1 mounts.

SETCLIENTID initialized a mutex in each clientid which was used by
some state management functions.  This patch adds initialization of
the mutex to EXCHANBE_ID, solving a crash in some state management
operations under NFSv4.1.