Change BUDDY_ADDR_T to void *
BUDDY_ADDR_T was originally defined to expand to caddr_t.  This has
the unfortunate effect that the pointer returned from Mem_Alloc must
be explicitly cast to be assigned to any pointer variable not of type
char * or caddr_t.  This has the effect of cluttering the code and
reducing type safety (if, for example, a function allocating objects
of a specific type were added somewhere, old code assigning its
pointers to something inappropriate would receive no warning because
of the cast.)