HashTable: Rewrite/Refactor
This patch accomplishes two goals.  First, it provides an
implementation of the *Ex series of hashtable calls as requested,
though the names are changed:

 * HashTable_GetLatch: This function looks up a key in a hash table
   and, on success and not-found, retains the lock on the hash
   partition.
 * HashTable_ReleaseLatched: This function releases a lock acquired in
   HashTable_GetLatch.
 * HashTable_SetLatched: This function sets a value in the hash table,
   the key must be the same as that supplied to HashTable_GetLatch.
 * HashTable_DelLatched: This function removes a value from the hash
   table.  The key supplied must be the same as that supplied to a
   previous call of HashTable_GetLatch.

Both HashTable_SetLatched and HashTable_DelLatched release the lock
held on the table.  See the Doxygen comments in HashTable.c for more
information.

HashTableGetLatch fills (and the -Latched functions consume) a datum
of type struct hash_latch.  It contains the index (identifying the
partition) and hash (identifying the place in the tree) values
computed during lookup and the node found (or NULL if no node was
found) so that the values need not be recomputed on later calls.  I
was asked about getting index value for a key, and it can be retrieved
from here.

The Set, Get, Del, GetRef, DelRef, and Test_And_Set functions were
implemented in terms of the above primitives.

The second goal was to cut down the number of allocations used in a
hash table and to give it better cache behaviour by merging the arrays
together into a single one of type struct hash_partition.