XFS FSAL - return correct file descriptor when opening directories
Fallback for handling directory opening in fsal_internal_handle2fd
was returning wrong file descriptors: the parenthesis were
placed incorrectly and was doing the equivalent of

fd = (open_by_handle() < 0);

so we were getting either 0 or 1 as file descriptor for a directory.
Combined with open/fchmod/close it was creating a land mine for the
open - a workload which feature a lot of directory making (like git clone)
was racing and opening unrelated files on FD 0 and 1 and using wrong
file descriptors for setting attributes. This explains why some permission
bits for some files were "morphing" between two tree scans by git status.

Signed-off-by: Max Matveev <makc@gmx.co.uk>
Signed-off-by: Jim Lieb <jlieb@panasas.com>