FSAL_check_quota - let's try to agree about its arguments
FSAL_check_quota is taking simple char pointer as first argument yet
COMMON_check_quota is stuck in the time warp and thinks it should
be fsal_path_t.

Because XFS and HPSS are the only FSALs which implement its own
check_quota methods and because of the other code fix COMMON_check_quota
to follow the new order.

Signed-off-by: Max Matveev <makc@gmx.co.uk>