PROXY FSAL - allow NFS v3 and NFS v2 without handle mapping
If the backend server handle can fit into the opaque part
of the protocol then there is no reason to not allow NFS v2 and
NFS v3 access via Proxy - in this case just use the handle
as is.

And make enable_handle_mapping independent of HANDLE_MAPPING
choice at the compile time to allow administator's control
over protocol selection.

Signed-off-by: Max Matveev <makc@gmx.co.uk>