Fix the rm -rf bug
cache_inode_new_entry takes an argument giving additional information
about a file: such as special device numbers, symlink content, and
whether the directory is newly created.  In this case, a new directory
created through ganesha is known to be empty and can be marked as such
in the cache.  cache_inode_readdir_populate had a single create_arg
used throughout any given invocation.  It was initialized to mark a
found (not created) directory at the beginning, but if a symlink was
encountered, the link content was copied into it.  As the create_arg
is a union and the content was never cleared, this caused directories
to be spuriously treated as new creates.

To fix it, we set the flag indicating a newly created directory to
false for every entry that is not a symbolic link.