Introduction of nfs4_sanity_check_FH and some very basic code refactoring for NFSv41.
Introducing nfs4_sanity_check_FH to the parts of the code where it was applicable.

Also some code refactoring. Removing some intermediate variables, used to store
the return value of some functions. In those cases, removing the variable
assignment from the 'if' statement. And some symmetry reorg for a better look :-)

Signed-off-by: Leonardo Tonetto <tonetto@linux.vnet.ibm.com>