PROXY FSAL - memory leak when dealing with names
fsal_internal_proxy_fsal_name_2_utf8 checks if len is 0 and allocates
memory. Callers of fsal_internal_proxy_fsal_name_2_utf8 do not set len
but provide a pointer for the string and assume that it's going to be
enough to avoid allocation. It isn't so on every call we leak the memory.

Signed-off-by: Max Matveev <makc@gmx.co.uk>