PROXY FSAL - don't clobber fd 0 on reconnect
During the initialization we forget to save the socket
in the thread context then on reconnect we close this socket
and try to connect again. And then another thread comes along
and does the same thing. And again. And again. As a result
we have several contexts using one socket - this is not going
to work too well.

To avoid problems like this move the code which deals with
creating RPC clients into separate function and call this
new function whenever you need a new RPC client.

Signed-off-by: Max Matveev <makc@gmx.co.uk>