Integrate cache inode rewrite and new api handling of entries
The big change here is a handle is now an object that gets allocated
and, of course, must be released.  These five functions is where it
all happens.

cache_inode_create calls the appropriate fsal method to create the object
in the undrelying filesystem.  This object is now passed to new entry.

cache_inode_new_entry takes a newly created fsal_obj_handle which
it adds to the new entry it creates.  Otherwise, in the case of an
already existing entry or on an error, it releases the handle itself.

cache_inode_get still takes a handle descriptor which it uses for the lookup.
If the lookup fails, it creates a handle object and passes it to new entry.

cache_inode_remove releases the fsal_obj_handle as part of its removal of
a cache entry.

Since the lookup of a root handle returns an object handle for it,
cache_inode_make_root now calls new entry directly.

Signed-off-by: Jim Lieb <jlieb@panasas.com>