FSAL_POSIX: fix permission check for non-owners in presence of FSAL_OWNER_OK
FSAL_OWNER_OK allows overriding owner permission checks if the file is
owned by the requester.  However, this bit should be cleared when it
isn't owned by the requester.

Without this, attempts to read a file not owned by the requester will
fail due to the missing_access bit being set.

Signed-off-by: David Flynn <davidf@rd.bbc.co.uk>