We have to find the export in a more reliable place.
Turns out pexport is not what it seems.  The cleanup of rpc handling
is deferred for now.  Right now, get the pointer from a more reliable
place.  This probably won't change if the cleanup happens because quota
testing (found bug but turning it on) is relative to the cache entry in
question anyway.

Signed-off-by: Jim Lieb <jlieb@panasas.com>