Added a new log function to print debug info during specific log levels.
The new function will trigger debug info like backtraces, global
variables, etc. to be printed after any log message that is less then or
equal to the log level of LOG_MESSAGE_DEBUGINFO, a new component.
LOG_MESSAGE_DEBUGINFO is set to NIV_NULL by default, so it has to be
explicitly enabled.

    Ported from 1.5.x:
    bf2652aac2c762ecde75133db85d74285b270f91

Signed-off-by: Jeremy Bongio <jbongio@linux.vnet.ibm.com>
Signed-off-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>