Fix ENOTDIR Blast failures
This patch corrects the blast test failures where in operations
appeared to fail with the ENOTDIR error code.  This error
occurred because when handle_size was zero, the PTFSAL_handlecmp
routine returns sucesful when it should not.  This patch modifies
the routine such that the handles are compared by the entire
OPENHANDLE_KEY_LEN bytes.

Signed-off-by: Allison Henderson <achender@linux.vnet.ibm.com>