Improve GPFS errors in building export context
Make the check if the GPFS file system exists more
stringent. Export path must actually be a sub-directory
under the GPFS mount point or the mount point itself.
Check for \0 or / after the portion of the export path
that is equal to the mount point. So, for example, if
mount point is /ibm/fs0:

/ibm/fs0 is a valid export path
/ibm/fs0/ is a valid export path
/ibm/fs01 is not a match
/ibm/fs0/dir is a valid export path

Cleanly report that the exported path doesn't exist.

Cleanly report that the exported path is not a directory.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>