Change API for the fs_loc_body_size && fs_da_addr_size vectors
Some MDSs have a const size layout or deviceinfo buffers, but
for some MDSs it is variable sized depending on the range requested
and the available space at client buffer. If the requested layout
(for example), is to big to fit into client's buffer, a smaller range
can be granted and the client will return again for the remaining
range.

What we do is pass the client's max-size into fs_loc_body_size()
&& fs_da_addr_size(), the MDS can then decide what it wants to
return.

While changing the API at fsal_api.h I have also fixed some
comments and typos in the pNFS API part.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Acked-by: Adam C. Emerson <aemerson@linuxbox.com>