Fix LOOKUPP to cross back over junctions into pseudo fs
LOOKUPP will check if the cache entry is the root of the
current export, and if so, will call
nfs4_op_lookupp_pseudo_by_exp to cross back into the pseudo
file system.

Note that the current export is used to determine which
pseudo fs entry is returned. This means that when duplicate
exports are supported again using different Pseudo paths for
each instance, LOOKUPP will return to the correct point in
the pseudo fs.

Note that this fix doesn't handle the case of junctions
inside an exported file system, but then LOOKUP doesn't handle
those either.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>