Change Behavior of fs_loc_body_size && fs_da_addr_size
Some MDSs have a const size layout or deviceinfo buffers, but
for some MDSs it is variable sized depending on the range requested
and the available space at client buffer. If the requested layout
(for example), is too big to fit into client's buffer, a smaller
range can be granted and the client will return again for the
remaining range.

In the variable size case the FSAL can return ~0UL in both of
these, and ganesha core will never allocate a buffer bigger
then the maximum requested by the client.

While changing the API at fsal_api.h I have also fixed some
comments and typos in the pNFS API part.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Acked-by: Adam C. Emerson <aemerson@linuxbox.com>