Add config block for log
Var names are the log component name (such as COMPONENT_FSAL).
The COMPONENT_ part is optional. COMPONENT_ is still required
in the environment.

Var values are the log levels (such as NIV_DEBUG). The NIV_
part is optional (and is now optional for environment and
SNMP set values.

Ignore case for components and level values (except component
must still be in all caps in environment).

Signed-off-by: Frank Filz <ffilz@us.ibm.com>