Add cache_inode_check_setattr_perms and need_wr_lock to cache_inode_lock_trust_attrs
Add a new permission checking routine in cache inode to validate permissions for
setattr.

cache_inode_check_setattr_perms will examine the attributes being set, and determine
what permission checks will need to be made by the FSAL. Since this is a complex
routine, and root is always allowed to set attr, the first check made is to exit
quickly if root (squashing has already been taken care of).

First, there is a general rule on owner and owner_group, a user can only set the
owner to himself, and owner_group to a group he is a member of.

owner and owner_group are may only be changed by the owner of the file, or a user
that has been granted WRITE_OWNER permission by ACL. Root/admin granting WRITE_OWNER
permission in an ACE for a specific user is used to implement "Take Owner" permission.
By using this, the admin grants a specific user the right to change the owner and
owner_group to himself (due to the general rule above).

mode and ACL may only be changed by the owner of a file, or a user granted WRITE_ACL
permission by an ACE.

size may be changed by anyone who has write permission to the file.

atime and mtime may be set to "now" by anyone who has write permission. Setting to
a specific time is only allowed for the owner of the file, or a user granted
WRITE_ATTR permissission in an ACL.

New setattr will require getting attr lock for write and we need to check trust
of attrs, so add ability in cache_inode_lock_trust_attrs to request write lock.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>