pnfs_panfs: conditionally enable pnfs_panfs at export creation time
Add a "pnfs_panfs" FS_specific option to enable/disable panfs
export per export-point.

This is so code can be always enabled, so not to bit rot

The FS_specific = ""; string was already parsed by Ganesha core
and is passed as the "const char *fs_options" param to ->create_export().

I defined a new fs_specific_has() function that parse a comma separated
list of key=value or just key, and returns if key was found, and optionally
it's value.
	(Example: FS_specific = "foo=baz,enable_A";)
This function is currently static to FSAL_VFS but could be made public if
needed elsewhere.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>