Restructure bitmap4 without heap allocation.
On 64 bit, the val+len structure is the same size as the bitmap
itself.  This is because 32 bits of padding is needed to align the
following pointer.  Combine the map and the length in to one struct and eliminate
the malloc and all the mem leak problems and frag'd heap.  On 32 bit, we
do not have the padding but eliminating the malloc and the pointer still yields
a single struct only 8 bytes bigger. Eliminating small malloc requests is also kinder
to the heap.

Signed-off-by: Jim Lieb <jlieb@panasas.com>