Allow passing SET_TO_SERVER_TIME to file system (LUSTRE, POSIX, VFS, XFS, GPFS)
VFS and XFS have a file descriptor, so are able to use utimesn which
provides nanosecond granularity and allows setting one or both of
atime and mtime to either a specified value or to current time. This
allows a "perfect" match with the SETATTR requests.

For LUSTRE and POSIX, we must use utimes which provides only microsecond
granularity and must set both times (though it can set both times to
current time). If only one time is set to server time, then we use
gettimeofday to set it. If one of the times is not set, then it
is "re-set" using the recently fetched attributes. This clearly is
not ideal, but is the best we can do.

GPFS OPENHANDLE_SET_XSTAT is able to handle all the intricacies of
SETATTR. Minor bug fix though, we were not originally setting
nseconds for atime and mtime...

Signed-off-by: Frank Filz <ffilz@us.ibm.com>