Workaround infrequently appearing Linux READDIR bug.
We observed Linux 3.4.4 send a new readdir, starting from the
offset of the last cookie in a prior, terminal READDIR reply
segment.  If Ganesha returned error in this case, Linux would
cache no dirents.

Adam's suggested hack is to return eod and no entries to such
a request.  This requires an extra AVL lookup to ensure we have
the last entry.  It's still fast, and goes away if Linux is
fixed.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>