Eliminate FSAL_NGROUPS_MAX, make alt_groups a pointer instead of fixed size array
In order to allow unlimited number of alt groups to the FSAL, we must
make the alt_groups array be a pointer instead of a fixed size array in
struct user_credentials.

The user_credentials alt_groups will normally just point to the
group list derived from the request credentials.

With that, blocking lock data must make a copy of the alt_groups since
the credentials attached to a blocked lock have a lifetime beyond the
life of the request.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>