make use of libnfsidmap easily configurable with new config parameter UseGetpwnam.
This is very useful where idmapd is not working or is not running
and one needs a temporary workaround. I made default properly set
based on whether one has libnfsidmap-support compiled in or not.
I.e. ganesha will throw a user-friendly error if you try disable
UseGetpwnam without libnfsidmap compiled in. The default behavior
is preserved if it *is* compiled in. So no one needs to change any
configuration.