Remove validate method from FSAL
Since we're not going to implement the cache control logic in this
version, I'd just as soon not have a completely un-used FSAL method
hanging around.  Besides, I think the prototype I had isn't what I
actually want, and it's easier to add a new method than change the
parameters of an existing one.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>