Fix cmake build of nodedb and posix fsal
nodedb has to build the 'marshal' program to process the nodedb.c
file.  Build this properly.

This archive must be linked to the POSIX fsal.  Create the archive
as PIC files so it can be linked onto the fsal .so

NOTE: this does not include a fix for automake builds.  POSIX is
broken in that case because the nodedb library is assumed to be a .so.
No, we are not going to do the autotools Fu to make it work.  The
whole point of the cmake port has been to not do such ill advised
acts.  Use cmake for POSIX builds.

Signed-off-by: Jim Lieb <jlieb@panasas.com>