Don't crash on illegal operations
There's a big, bad per-worker stat array that tracks calls, errors,
successes, and so forth for every NFSv4 operation.  It now checks that
the operation number it's tallying is in the array before scribbling
to bogus memory.  This function should almost certainly be removed as
part of the new stats/performance work underway.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>