fsal_pnfs: Default variable recall_cookies[] array to one (was 0)
With pnfs_panfs we have the structure fsal_layoutreturn_arg as part of an
IOCTL, which must have a fixed size. (IOCTL rules)

This is not a problem for Panfs since it will only have at most one recall
per layout_get. So a single entry will always be enough.

This patch does not changes the API in any way and or the possibilities
of having multiple recalls per return. It only defaults the array size to
be minimum 1 instead of 0.

Sign-off-by: Boaz Harrosh <bharrosh@panasas.com>