FSAL VFS - deal with XFS-specific handles
Abstract operations which deal with creating and
manipulating of handles and add XFS-specific
handle processing.

I'm still using vfs_file_handle_t as container
for XFS handles but every place which used
to call vfs_XXX_by_handle_at is now going via
indirection vector in vfs_fsal_export.

The vector has to be in the export because we
need to be able to create vfs_fsal_obj_handle
out of the opaque file handle which could have been
cached on the client.

Signed-off-by: Max Matveev <makc@gmx.co.uk>