gitignore: Add build*/
It looks like everyone is costumed to build_foo naming convention
for building with cmake. So It should be safe to make this a policy:

	NAME your compilation directory starting with build*
	and it will be ignored by git..

I know that the strict among us would say that it needs to be
a folder outside of ganesha's git. But for developers this means
inconvenience  because it means we need everything one directory
deeper, in an already deep directory structure. So I would like to
keep everything under ~/git/ganesha/... thank you.

personally I don't mine the build* naming convention, though some
people might prefer:
	.build*
With the "dot" it should be more of a private "ignored" stuff.

So please complain about this patch if you don't like it or
Jim can take it.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>