9P/RDMA: Protect send buffers against early re-use
For 9P/RDMA, ganesha has two sets of buffers : recv and send buffers.
Each recv buffer is associated to a send buffer.
When a request is received on a recv buffer, it is processed and the associated
send buffer is used to send the reply.

But the current code fails to check that the previous message that was to be
sent using this send buffer has actually been sent and that this buffer can
be reused. In case of high traffic, this leads to buffer overwriting and
duplicate messages.

This patch adds a mutex to each buffer. Before processing a request, the mutex
is taken, and it will be released after the reply has been sent. If another request
wants to use the send buffer, it will have to wait for the message to be really
sent.

Signed-off-by: Simon Derr <simon.derr@bull.net>
(cherry picked from commit 859bca0298ebc7010bf11ee511ec90c2719e0ddc)

Signed-off-by: Jim Lieb <jlieb@panasas.com>