Fix USE_DBUS-OFF build problems
Not enough of the dbus interface was ifdef'd out...
This leaves the client manager and export manager infrastructure
untouched so it is available for other uses.

Note: stats are still gathered, you just can't get them out without
dbus.  Also, an ifdef in nfs4_Compound.c was just commented out.  This
is left as a marker until the old stats code has been removed.

Signed-off-by: Jim Lieb <jlieb@panasas.com>