Remove current volatile filehandle implementation
The current volatile filehandle implementation does not work well.  It
expires all filehandles on server reboot, which is of dubious value.
There's also facility to let the FSAL set an expire type that isn't
used for anything.

If we do decide we want volatile filehandle support, we could get it
later by letting the FSAL return an EXPIRED error that is propagated
to the user and exposing an expire type as a per-filesystem attribute.
(An upcall to let it blow out an expired handle might be nice, too.)

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>