Close the open fd where we can before unlinking the file.
This patch addresses the issue on nfs-ganesha running in Panasas environment
where unlinking an open file dscriptor results into silly rename of the file on
underlying filesystem. Intention here is to close the fd iff its the only open
reference, not pinned and no clients have locks on the said fd.

Signed-off-by: Sachin Bhamare <sbhamare@panasas.com>