Fix MOUNTED_ON_FILEID, setting from export if entry is export root
Any cache inode routine that returns attributes must check
if the entry is root inode of current export, and if so,
grab the exp_mounted_on_fileid from that export. For
readdir, the actual cache inode entry attributes are changed
because a separate set of attributes are not available. This
will be ok since the attributes are immediately consumed
during the readdir callback.

Since the current export is used, this will result in the
correct mounted_on_fileid when duplicate exports are again
supported.

Signed-off-by: Frank Filz <ffilz@us.ibm.com>

Manual merge from 6aa4a7800602b2f89

Signed-off-by: Jim Lieb <jlieb@panasas.com>