Use allowed/denied return from cache_inode_access_sw to simplify v3/4 ACCESS
By using the ability of cache_inode_access_sw to return the set of
allowed bits combined with the ability to ask for ACL evaluation to not
terminate on an applicable DENY ACE, we can make the access determination
for ACCESS with a single permission evaluation.

Also, provide a shared routine for performing the ACCESS logic.

Signed-off-by: Frank S. Filz <ffilz@us.ibm.com>