Add pNFS flags to owner and use a better hash
RFC 5661 says that servers are allowed (but not required) to treat
identical co_owner values sent with different pNFS flags in
EXCHANGE_ID as describing different owners.  The Linux client,
however, responds to being given a known clientid when connecting to a
DS that is colocated on an MDS by calling DESTROY_CLIENTID and
re-exchanging.

This is a particularly aggressive way not to support trunking.

Also switch to using CityHash for co_owner values rather than a
moddy-loopy-sum over all the bytes.

Signed-off-by: Adam C. Emerson <aemerson@linuxbox.com>