Use fsal_set_credentials in GPFSFSAL_unlink, GPFSFSAL_rename, and GPFSFSAL_link
Due to atomicity problems, GPFS needs to do the permission checking for
directory mutating operations, so set the thread credentials to allow that and
strip the old explicit permission checking.

Ported from 1.5 93ed49514aa8984b7c7aa1c197e18cfdda912ab1

This is a slightly different set of routines, but the net effect is the same.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>