Handle errors in PseudoFS READDIR and LOOKUP gracefully
If there is a problem with a junction, we should handle it gracefully,
not bomb out the whole READDIR, In this case, we skip the entry.

This could occur because the exported directory has been removed or
the attributes could not be fetched. There are permissions fixes that
will prevent the ACE4_READ_ATTR from blocking reading attributes
though ACE4_READ_ACL could still come into play if a client did a
READDIR asking for ACLs (which I don't expect will ever be done).

Also, return more sensible errors in other places.

Signed-off-by: Frank S. Filz <ffilz@us.ibm.com>