Allow v3 to return fileid on READDIR in 444 directory
When a user only has read access in a directory, the client is still
dependent on getting the fileid for each file. If it does not, the
client fails to show any directory entries (other than . and ..).

Unfortunately, for Linux clients, this only works for v3 (other
clients are unknown). v4 Linux client must get all the attributes
or it does not work correctly.

There was also a bug where Ganesha was actually not returning all
the entries.

Signed-off-by: Frank S. Filz <ffilz@us.ibm.com>