Fix a problem uncovered by RNM20
Rename of file to a hard link to the file is directed by the RFCs to
do nothing. Code was not correctly handling and causing the "from"
name to be removed from Ganesha's directory cache.

The scenario in question is caused by the following sequence:

echo foo > /mnt4/file
ln /mnt4/file /mnt4/link
mv /mnt4/file /mnt4/link

It would appear that that should result in just one hard link to
the inode, named /mnt4/link, however, the RFCs specify that this
is a no-op operation, which means the file remains with two hard
links.

Signed-off-by: Frank S. Filz <ffilz@us.ibm.com>