Switch credentials for ops in VFS fsal.
Switch creds to client user for ops that morph the directory.
This is done to manage creds and in the case of rename deal with
races.

Left off are makenode which needs the "make safe" for block, char etc.
Since these don't add anything other than an inode, work around.  You
need caps privs in order to create these things anyway (and owned by root).
The other makenode types are ephemeral anyway so why bother.

Signed-off-by: Jim Lieb <jlieb@panasas.com>