Revamp NFS v4 handle to eliminate xattr, pseudofs, and replace ds_flag with generic flags
With the old xattr implementation gone, and PseudoFS not needing special fields anymore,
all that remains is ds_flag. We replace all the special fields in the handle with a simple
flags field and define a value for DS handles. A future xattr implementation will use a
different flag value, and will use fsopaque to stash it's magic in.

Signed-off-by: Jeremy Bongio <jbongio@us.ibm.com>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>