Changes to pseudofs code for new nfsv4 file handle.
The PseudoFS had a problem where if Ganesha was restarted with a
different set of exports (exports added or removed), the handle for a
particular PseudoFS node could change. This would cause disruption to
any NFS v4 client that had a handle to a PseudoFS node cached.

To resolve this problem, we make the PseudoFS be exportid 0, and then
instead of the PseudoFS handle being a 16 bit number which was an array
index, the handle now includes the following in the fsopaque portion:

uint64_t Cityhash64 hash of the path to the PseudoFS entry
uint16_t length of the path
char[]   portion of the path that fits into available portion of handle

In the future, FSAL_PSEUDO is planned that will build the same handle
but otherwise will work as a normal FSAL and special PseudoFS code will
be history.

Signed-off-by: Jeremy Bongio <jbongio@us.ibm.com>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>