CID#18258 Structurally dead code in str_lockt
This instruction could never be reached under any circumstances.
As suggested by Frank Filz, let's keep the ending
return statement and remove the default case.
This is to avoid missing case detection with
strict compile can keep the code honest. We have
had number of bugs happen because someone added
to an enum and didn't add cases to every switch on
that enum.