Fix use of current stateid
NFS v4.1 Adds a new special stateid: other=0, seqid=1. This stateid indicates
to use the "current stateid" from the last operation that returned a stateid.

There were two problems, first, CLOSE did not request to allow that stateid
because it was using the wrong flag constants.

Second, when we found a valid current stateid, we did not escape to check
it properly.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>